requirements.txt
fastapi
uvicorn
pydantic
pymongo
requests
python-dotenv
openai